fix functions from different versions 2
