+++ core/src/main/spark_3.1_mount/scala/doric/syntax/All.scala
@@ -21,10 +21,10 @@ private[doric] trait All
     with AggregationColumns31
     with BooleanColumns31
     with NumericColumns31
+    with NumericColumns2_31
     with StringColumns31
     with BinaryColumns30_31
     with ArrayColumns3x
     with CommonColumns3x
     with MapColumns3x
     with StringColumn3x
-    with NumericColumns2_31
